requirements.txt
fastapi
uvicorn
python-socketio
google-genai
python-dotenv
requests
